+++ sandbox/measure_reportread.py
@@ -1,8 +1,8 @@
 import time
 
 from secfsdstools.a_config.configmgt import Configuration
-from secfsdstools.d_index.indexdataaccess import IndexReport
-from secfsdstools.d_index.indexing import BaseReportIndexer
+from secfsdstools.c_index.indexdataaccess import IndexReport
+from secfsdstools.c_index.indexing import BaseReportIndexer
 from secfsdstools.e_read.basereportreading import NUM_TXT, PRE_TXT, SUB_TXT
 from secfsdstools.e_read.companycollecting import CompanyReportCollector
 from secfsdstools.e_read.reportreading import ReportReader