+++ secfsdstools/c_download/secdownloading.py
@@ -22,8 +22,10 @@ class SecZipDownloader(BaseDownloader):
     table_re = re.compile('<TABLE.*?>.*</TABLE>', re.IGNORECASE + re.MULTILINE + re.DOTALL)
     href_re = re.compile("href=\".*?\"", re.IGNORECASE + re.MULTILINE + re.DOTALL)
 
-    def __init__(self, zip_dir: str, urldownloader: UrlDownloader, execute_serial: bool = False):
+    def __init__(self, zip_dir: str, parquet_dir: str,
+                 urldownloader: UrlDownloader, execute_serial: bool = False):
         super().__init__(zip_dir=zip_dir, urldownloader=urldownloader,
+                         parquet_dir_typed=os.path.join(parquet_dir, 'quarter'),
                          execute_serial=execute_serial)
 
     @classmethod
@@ -43,6 +45,7 @@ class SecZipDownloader(BaseDownloader):
 
         urldownloader = UrlDownloader(user_agent=configuration.user_agent_email)
         return SecZipDownloader(zip_dir=configuration.download_dir,
+                                parquet_dir=configuration.parquet_dir,
                                 urldownloader=urldownloader)
 
     def _get_available_zips(self) -> List[Tuple[str, str]]:
@@ -50,11 +53,16 @@ class SecZipDownloader(BaseDownloader):
         first_table = self.table_re.findall(content.text)[0]
         hrefs = self.href_re.findall(first_table)
 
-        hrefs = ['https://www.sec.gov' + href[6:-1] for href in hrefs]
+        hrefs = [f'https://www.sec.gov{href[6:-1]}' for href in hrefs]
         return [(os.path.basename(href), href) for href in hrefs]
 
     def _calculate_missing_zips(self) -> List[Tuple[str, str]]:
         dld_zip_files = self._get_downloaded_zips()
+        transformed_parquet = self._get_transformed_parquet()
         zips_to_dld_dict = self._get_available_zips()
 
-        return [(name, href) for name, href in zips_to_dld_dict if name not in dld_zip_files]
+        # define which zip files don't have to be downloaded
+        download_or_transformed_zips = set(dld_zip_files).union(set(transformed_parquet))
+
+        return [(name, href) for name, href in zips_to_dld_dict if
+                name not in download_or_transformed_zips]