+++ secfsdstools/e_read/zipreportreading.py
@@ -9,7 +9,7 @@ import pandas as pd
 
 from secfsdstools.a_config.configmgt import Configuration, ConfigurationManager
 from secfsdstools.a_utils.fileutils import read_df_from_file_in_zip
-from secfsdstools.d_index.indexdataaccess import create_index_accessor
+from secfsdstools.c_index.indexdataaccess import create_index_accessor
 from secfsdstools.e_read.basereportreading import BaseReportReader