+++ tests/e_read/test_companyreading.py
@@ -2,7 +2,7 @@ import os
 from unittest.mock import MagicMock
 
 from secfsdstools.a_config.configmgt import Configuration
-from secfsdstools.d_index.indexdataaccess import IndexReport
+from secfsdstools.c_index.indexdataaccess import IndexReport
 from secfsdstools.e_read.companyreading import CompanyReader
 
 CURRENT_DIR, _ = os.path.split(__file__)