+++ tests/e_read/test_zipreportreading.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 import pytest
 
 from secfsdstools.a_config.configmgt import Configuration
-from secfsdstools.d_index.indexdataaccess import IndexFileProcessingState
+from secfsdstools.c_index.indexdataaccess import IndexFileProcessingState
 from secfsdstools.e_read.zipreportreading import ZipReportReader
 
 CURRENT_DIR, _ = os.path.split(__file__)
@@ -54,7 +54,7 @@ def test_cm_get_zip_by_name():
     instance = IndexFileProcessingState(fileName="", status="", entries=0, processTime="",
                                         fullPath=PATH_TO_ZIP)
 
-    with patch("secfsdstools.d_index.indexdataaccess.DBIndexingAccessor.read_index_file_for_filename",
+    with patch("secfsdstools.c_index.indexdataaccess.DBIndexingAccessor.read_index_file_for_filename",
                return_value=instance):
         zipreader = ZipReportReader.get_zip_by_name(name="2010q1.zip",
                                                     configuration=Configuration(db_dir="",